requirements.txt
fastapi==0.106.0
uvicorn==0.23.2
sentence-transformers==2.2.2
psycopg2-binary==2.9.9
numpy==1.24.4
requests==2.32.3
PyPDF2==3.0.1
pre-commit==3.1.0
flake8==6.1.0
